Rename CommunitiesProvider
