+++ providers/index.js
@@ -1,7 +1,7 @@
 import AppProvider from './AppProvider';
 import AuthProvider from './AuthProvider';
 import CampusesProvider from './CampusesProvider';
-import CommunityProvider from './CommunityProvider';
+import CommunitiesProvider from './CommunitiesProvider';
 import ContentFeedProvider from './ContentFeedProvider';
 import ContentItemProvider from './ContentItemProvider';
 import CurrentPersonProvider from './CurrentPersonProvider';
@@ -9,7 +9,7 @@ import CurrentUserProvider from './CurrentUserProvider';
 import EventProvider from './EventProvider';
 import EventsProvider from './EventsProvider';
 import FeatureFeedProvider from './FeatureFeedProvider';
-import FeatureProvider from './FeatureProvider'
+import FeatureProvider from './FeatureProvider';
 import FeedFeaturesProvider from './FeedFeaturesProvider';
 import GroupProvider from './GroupProvider';
 import GroupsProvider from './GroupsProvider';
@@ -21,7 +21,7 @@ export {
   AppProvider,
   AuthProvider,
   CampusesProvider,
-  CommunityProvider,
+  CommunitiesProvider,
   ContentFeedProvider,
   ContentItemProvider,
   CurrentPersonProvider,
